+++ src/pages/Folder/index.jsx
@@ -4,7 +4,7 @@ import FolderHeader from "./components/FolderHeader";
 import FolderAddLinkArea from "./components/FolderAddLinkArea";
 import FolderCategory from "./components/FolderCategory";
 import FolderControl from "./components/FolderControl";
-import styles from "./Folder.module.css";
+import styles from "./styles.module.css";
 import SearchIcon from "assets/Search.png";
 import { useEffect, useState } from "react";
 import { getFolderList, getLinks } from "api/api";
