+++ src/pages/Share/index.jsx
@@ -1,7 +1,7 @@
 import Header from "components/Header";
 import Footer from "components/Footer";
 import CardList from "components/CardList";
-import styles from "./Share.module.css";
+import styles from "./styles.module.css";
 import SearchIcon from "assets/Search.png";
 import { useEffect, useState } from "react";
 import { getFolderInfo } from "api/api";
